rewrite some logic
While trying to replace arrays with classes I found a few issues
that make no sense or little sense to me.
For example, the original code would use the title of the
milestone as key to the array, but this way if two milestones have
the same title (quite possible with semver and different
repositories), one will replace the other.
Also I found that arrays were quite abused, IMHO.
Finally I have removed properties that are not used in the view,
and, in the scope of this exercise, don't make much sense to have
around.